requirements.txt
fastapi~=0.104.1
sqlalchemy~=2.0.23
pydantic-settings~=2.1.0
pydantic~=2.5.2
python-jose~=3.3.0
passlib~=1.7.4
httpx~=0.28.1
uvicorn~=0.24.0